# api/swagger/swagger.yaml
basePath: /api
definitions:
...
  dto.CompleteCycleRequest:
    properties:
      move_to_cycle_id:
        type: string
    type: object
...
  dto.CreateCycleRequest:
    properties:
      ends_at:
        type: string
      goal:
        type: string
      name:
        type: string
      starts_at:
        type: string
    required:
    - ends_at
    - starts_at
    type: object
...
  dto.CycleResponse:
    properties:
      completed_task_count:
        type: integer
      created_at:
        type: string
      created_by:
        type: string
      ends_at:
        type: string
      goal:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      is_completed:
        type: boolean
      name:
        type: string
      project_id:
        type: string
      starts_at:
        type: string
      task_count:
        type: integer
      updated_at:
        type: string
    type: object
...
  dto.UpdateCycleRequest:
    properties:
      ends_at:
        type: string
      goal:
        type: string
      is_completed:
        type: boolean
      name:
        type: string
      starts_at:
        type: string
    type: object
...
  transport.ErrorResponse:
    properties:
      error:
        properties:
          code:
            type: string
          message:
            type: string
        type: object
    type: object
...
info:
  contact: {}
  description: Self-hosted project management tool API
  title: Plume API
  version: "1.0"
paths:
...
  /projects/{id}/cycles:
    get:
      description: Returns all cycles for a project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.CycleResponse'
            type: array
      summary: List cycles
      tags:
      - cycles
    post:
      consumes:
      - application/json
      description: Creates a new cycle in a project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Cycle details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCycleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CycleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/transport.ErrorResponse'
      summary: Create a cycle
      tags:
      - cycles
  /projects/{id}/cycles/{cycleId}:
    delete:
      description: Permanently deletes a cycle
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Cycle ID
        in: path
        name: cycleId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/transport.ErrorResponse'
      summary: Delete a cycle
      tags:
      - cycles
    put:
      consumes:
      - application/json
      description: Updates cycle fields
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Cycle ID
        in: path
        name: cycleId
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCycleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CycleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/transport.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/transport.ErrorResponse'
      summary: Update a cycle
      tags:
      - cycles
  /projects/{id}/cycles/{cycleId}/activate:
    post:
      description: Sets a cycle as the active cycle for its project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Cycle ID
        in: path
        name: cycleId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CycleResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/transport.ErrorResponse'
      summary: Activate a cycle
      tags:
      - cycles
  /projects/{id}/cycles/{cycleId}/complete:
    post:
      consumes:
      - application/json
      description: Marks a cycle as completed and handles incomplete tasks
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Cycle ID
        in: path
        name: cycleId
        required: true
        type: string
      - description: Completion options
        in: body
        name: body
        schema:
          $ref: '#/definitions/dto.CompleteCycleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CycleResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/transport.ErrorResponse'
      summary: Complete a cycle
      tags:
      - cycles
  /projects/{id}/cycles/active:
    get:
      description: Returns the active cycle for a project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CycleResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/transport.ErrorResponse'
      summary: Get active cycle
      tags:
      - cycles
...
  /projects/{id}/tasks/export:
    get:
      description: Exports all tasks in a project as CSV or JSON
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: 'Export format: csv or json (default json)'
        in: query
        name: format
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Export project tasks
      tags:
      - export
...
  /projects/{id}/time-entries/export:
    get:
      description: Exports time entries for a project's tasks as CSV or JSON
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: 'Export format: csv or json (default json)'
        in: query
        name: format
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Export time entries
      tags:
      - export
...
swagger: "2.0"